requirements.txt
# Text Similarity + RAG - Requirements
# PyTorch는 CUDA 버전에 맞게 먼저 별도 설치
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126

# FastAPI Backend
fastapi
uvicorn[standard]
python-multipart
pydantic

# HuggingFace
transformers
accelerate
sentence-transformers
sentencepiece

# ML/Computation
numpy
scikit-learn

# RAG / Vector Store
chromadb
pypdf
python-docx

# Utils
python-dotenv